test(swoole): Fix HttpKernelHttpDriver test
Configuration of Request's trusted hosts and proxies is no longer there but in separate bootable
class
